feat: multi-native language picker in StudentForm and OnboardingStep2
Replace single-select native language dropdown with MultiSelect component
supporting 1-5 languages as chips. Extract MultiSelect to shared component.
Fix data loss bug where editing a student with multiple native languages
would silently drop all but the first. Closes #630
